Refactor the files
