+++ src/main/java/example/cashcard/adapter/persistance/CashCardRepository.java
@@ -1,11 +1,12 @@
-package example.cashcard;
+package example.cashcard.adapter.persistance;
 
+import example.cashcard.domain.cashCard.CashCard;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-interface CashCardRepository extends CrudRepository<CashCard, Long>, PagingAndSortingRepository<CashCard, Long> {
+public interface CashCardRepository extends CrudRepository<CashCard, Long>, PagingAndSortingRepository<CashCard, Long> {
     CashCard findByIdAndOwner(Long id, String owner);
     Page<CashCard> findByOwner(String owner, PageRequest pageRequest);
     boolean existsByIdAndOwner(Long id, String owner);
